requirements.txt
langchain
langchain_community
python-dotenv
ipykernel
langchain-text-splitters
langchain-openai
sentence-transformers
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
sqlAlchemy
mysql-connector-python
sentence-transformers
langchain_huggingface
faiss-cpu
pypdf
PyMuPDF
python-multipart
langchain_google_vertexai
streamlit